pyproject.toml
[project]
name = "wsp-balsa"
description = "Python tools for travel demand forecasting applications and analyses"
dynamic = ["version"]
readme = "README.md"
authors = [
    { name = "Brian Cheung", email = "brian.cheung@wsp.com" }
]
requires-python = ">=3.9,<3.14"
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: BSD License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3"
]
dependencies = [
    "pandas>=0.24",
    "numpy>=1.21",
    "numexpr>=2.6",
]

[project.optional-dependencies]
plotting = [
    "matplotlib>=3.0",
]
matrices = [
    "numba>=0.35",
    "openmatrix",
]
quarto = [
    "pyyaml>=6.0.3",
    "quarto-cli>=1.8.27",
]
tables = [
    "h5py>=3.8.0",
]

[dependency-groups]
dev = [
    "build>=1.1.1",
    "ipykernel>=6.16.2",
    "pydata-sphinx-theme>=0.13.3",
    "pytest>=7.4.4",
    "sphinx>=5.3.0",
    "sphinx-favicon>=1.0.1",
    "twine>=4.0.2",
]

[project.urls]
Homepage = "https://github.com/wsp-sag/balsa"
Issues = "https://github.com/wsp-sag/balsa/issues"

[build-system]
requires = ["setuptools>=64", "setuptools-scm>=8"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools_scm]
version_file = "src/wsp_balsa/_version.py"

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
]
pythonpath = "src"

[tool.ruff]
line-length = 120
target-version = "py311"
